pyproject.toml
[build-system]
requires = ['setuptools>=82.0.1', 'wheel>=0.46.3', 'setuptools_scm[toml]>=9.2.2']
build-backend = 'setuptools.build_meta'

[tool.uv]
required-version = "0.10.9"

[project]
name = "baseapp_backend"
dynamic = ["version"]
description = "BaseApp"
readme = "README.md"
requires-python = ">=3.11"
license = "BSD-3-Clause"
authors = [
    { name = "The SilverLogic", email = "dev@tsl.io" }
]
urls = { Homepage = "https://github.com/silverlogic/baseapp-backend" }
dependencies = [
    "Django>=4.2.15",
    "daphne>=4.0",
    "channels>=4.0",
    "django-filter>=23.3",
    "celery>=5.2",
    "django-pgtrigger==4.17.0",
    "swapper>=1.3",
    "django-model-utils>=4.3",
    "djangorestframework>=3.14,<3.17",
    "drf-nested-routers>=0.93.5",
    "djangorestframework-expander>=0.2",
    "drf-extra-fields>=3.5",
    "psycopg2-binary>=2.9",
    "django-phonenumber-field>=7.0",
    "phonenumbers>=8.13",
    "easy_thumbnails>=2.8",
    "djmail>=2.0",
    "django-constance>=4.3.2",
    "django_jinja>=2.11",
    "django-pghistory==3.9.1",
    "django-ipware>=7.0.1",
    "python-ipware>=3.0.0",
    "requests>=2.32.3",
    "stevedore>=5.7.0",
    "pydantic>=2.12.5",
    "drf-spectacular==0.29.0",
]

[project.optional-dependencies]
auth = [
    "baseapp-django-trench>=0.4.4",
    "djangorestframework-simplejwt[crypto]>=5.2,<5.6",
]
referrals = ["hashids==1.3.1"]
graphql = [
    "graphene>=3.2",
    "graphene-django>=3.2.0",
    "graphene_django_optimizer>=0.10",
    "django-graphql-geojson>=0.1.4",
    "graphene-django-query-optimizer>=0.10.15",
    "graphene-file-upload==1.3.0",
    "baseapp-django-channels-graphql-ws==1.0.0rc9",
]
pages = [
    "django-quill-editor>=0.1",
    "django-translated-fields>=0.13",
]
payments = ["stripe>=11.5.0"]
notifications = [
    "django-notifications-hq>=1.8",
    "firebase-admin>=6.5.0",
    "django-push-notifications[WP,apns-async,FCM]>=3.2",
]
urlshortening = ["short_url>=1.2.2"]
messagetemplates = [
    "sendgrid>=6.9",
    "django-sendgrid-v5>=1.2",
    "beautifulsoup4>=4.6",
    "django-nested-admin>=3.3",
    "lxml>=4.9",
    "nh3>=0.2",
    "django-ckeditor>=6.5",
    "django-prose-editor[sanitize]>=0.24.0",
]
wagtail = [
    "wagtail>=7.0",
    "wagtail-headless-preview==0.8",
]
socialauth = [
    "hashids==1.3.1",
    "rest-social-auth>=9.0.0",
    "social-auth-core>=4.5.4",
    "social-auth-app-django==5.6.0",
]
pdf = ["pypdf>=6,<7"]

[dependency-groups]
dev = [
    "pytest==7.4.2",
    "pytest-cov==4.1.0",
    "pytest-django==4.5.2",
    "pytest-asyncio==0.21.1",
    "pytest-celery==1.2.1",
    "freezegun==1.2.1",
    "responses==0.25.3",
    "factory-boy==3.2.1",
    "black==26.3.1",
    "flake8==7.0.0",
    "isort==5.13.2",
    "channels-redis==4.0.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt[crypto]==5.5.1",
    "drf-nested-routers==0.93.5",
    "easy-thumbnails==2.10.1",
    "rest-social-auth==9.0.0",
    "urllib3==2.6.3",
    "channels==4.0.0",
    "django-notifications-hq",
    "django-push-notifications[apns-async,fcm,wp]==3.3.0",
    "baseapp-backend[auth,graphql,messagetemplates,notifications,pages,payments,pdf,referrals,socialauth,urlshortening,wagtail]",
    "testproject",
]

[tool.uv.sources]
django-notifications-hq = { git = "https://github.com/openwisp/django-notifications.git", rev = "f4918c2409a734ec1aa9798d86c3a729b505f753" }
baseapp-backend = { workspace = true }
testproject = { path = "testproject", editable = true }

[tool.setuptools_scm]
fallback_version = "0.0.0.dev"
tag_regex = "^baseapp@v(?P<version>[vV]?\\d+(?:\\.\\d+){0,2}[^\\+]*)(?:\\+.*)?$"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
baseapp = ["*.j2", "*.html", "*.png"]
baseapp_core = ["*.j2", "*.html", "*.png"]
baseapp_reactions = ["*.j2", "*.html", "*.png"]
baseapp_reports = ["*.j2", "*.html", "*.png"]
baseapp_ratings = ["*.j2", "*.html", "*.png"]
baseapp_follows = ["*.j2", "*.html", "*.png"]
baseapp_blocks = ["*.j2", "*.html", "*.png"]
baseapp_pages = ["*.j2", "*.html", "*.png"]
baseapp_payments = ["*.j2", "*.html", "*.png"]
baseapp_notifications = ["*.j2", "*.html", "*.png"]
baseapp_comments = ["*.j2", "*.html", "*.png"]
baseapp_cloudflare_stream_field = ["*.html", "*.js", "*.map"]
baseapp_url_shortening = ["*.html", "*.js", "*.map"]
baseapp_organizations = ["*.j2", "*.html", "*.png"]
baseapp_message_templates = ["*.j2", "*.html", "*.png"]
baseapp_chats = ["*.j2", "*.html", "*.png"]
baseapp_e2e = ["*.j2", "*.html", "*.png"]
baseapp_drf_view_action_permissions = ["*.j2", "*.html", "*.png"]
baseapp_social_auth = ["*.j2", "*.html", "*.png"]
baseapp_wagtail = ["*.j2", "*.html", "*.png"]
baseapp_profiles = ["*.j2", "*.html", "*.png", "*.jpg"]
baseapp_auth = ["*.j2", "*.html", "*.png"]
baseapp_referrals = ["*.j2", "*.html", "*.png"]
baseapp_pdf = ["*.j2", "*.html", "*.png"]
baseapp_api_key = ["*.j2", "*.html"]

[project.entry-points."baseapp.plugins"]
baseapp_auth = "baseapp_auth.plugin:AuthPlugin"
baseapp_profiles = "baseapp_profiles.plugin:ProfilesPlugin"
baseapp_pages = "baseapp_pages.plugin:PagesPlugin"
baseapp_organizations = "baseapp_organizations.plugin:OrganizationsPlugin"
baseapp_comments = "baseapp_comments.plugin:CommentsPlugin"
baseapp_payments = "baseapp_payments.plugin:PaymentsPlugin"
baseapp_referrals = "baseapp_referrals.plugin:ReferralsPlugin"
baseapp_pdf = "baseapp_pdf.plugin:PDFPlugin"
baseapp_message_templates = "baseapp_message_templates.plugin:MessageTemplatesPlugin"
baseapp_claudflare_stream_field = "baseapp_cloudflare_stream_field.plugin:CloudflareStreamFieldPlugin"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
